Emphasis: flipped board, take 1
Not entirely happy with the way this is going, might start again.

Basically I messed up the orientation of the main board relative to
the control board: to avoid having the patch_sm stick out the side,
parts need to be mounted on the "back" of main board. But this flips
the orientation of patch_sm and ICs, so everything needs to be
re-routed.